Move `git/common` to `git`
